🆙 remove warning issue

+++ src/Unit_MiniEncoderC.cpp
@@ -102,7 +102,7 @@ uint8_t UNIT_MINIENCODERC::getI2CAddress(void) {
 
     uint8_t RegValue;
 
-    _wire->requestFrom(_addr, 1);
+    _wire->requestFrom(_addr, (uint8_t)1);
     RegValue = Wire.read();
     return RegValue;
 }
@@ -114,7 +114,7 @@ uint8_t UNIT_MINIENCODERC::getFirmwareVersion(void) {
 
     uint8_t RegValue;
 
-    _wire->requestFrom(_addr, 1);
+    _wire->requestFrom(_addr, (uint8_t)1);
     RegValue = Wire.read();
     return RegValue;
 }
